Splitted HF text completion and embedding generation implementations
